lint + auth check

+++ controller/playlistController.js
@@ -5,14 +5,8 @@ exports.getAll = (req,res) =>
 {
     connection.query("SELECT `id` FROM `playlists` ORDER BY `playlists`.`created_at` DESC",(error,rows,fields)=>
     {
-        if (error)
-        {
-            response.status(400,error,res);
-        }
-        else
-        {
-            response.status(200,rows,res);
-        }
+        if (error) return response.status(400,error,res);
+        return response.status(200,rows,res);
     })
 }
 
@@ -20,52 +14,31 @@ exports.getByID = (req,res) =>
 {
     connection.query("SELECT * FROM `view_playlist` WHERE `id` = ?",[req.params.id],(error,rows,fields)=>
     {
-        if (error)
+        if (error) return response.status(400,error,res);
+        if (rows.length<1) return response.status(404,'Playlist not found',res);
+
+        const row = rows[0];
+        row.artists = [];
+        connection.query('SELECT `login`,`pseudoname`,`username` FROM `view_playlist_artists` WHERE `playlistID` = ? ORDER BY `view_playlist_artists`.`artistPlaylistPosition`',[row.id],(error,artists)=>
         {
-            response.status(400,error,res);
-        }
-        else if (rows.length<1)
-        {
-            response.status(404,'Playlist not found',res);
-        }
-        else
-        {
-            const row = rows[0];
-            row.artists = [];
-            connection.query('SELECT `login`,`pseudoname`,`username` FROM `view_playlist_artists` WHERE `playlistID` = ? ORDER BY `view_playlist_artists`.`artistPlaylistPosition`',[row.id],(error,artists)=>
+            if (error) return response.status(400,error,res);
+            artists.forEach(artist=>
             {
-                if (error)
+                row.artists.push({login:artist.login,name:artist.pseudoname?artist.pseudoname:artist.username});
+            });
+            if (req.user)
+                connection.query('SELECT `id` FROM `playlist_likes` WHERE `playlistID` = ? AND `userLogin` = ?',[row.id,req.user.login],(error,likes)=>
                 {
-                    response.status(400,error,res);
-                }
-                else
-                {
-                    artists.forEach(artist=>
-                    {
-                        row.artists.push({login:artist.login,name:artist.pseudoname?artist.pseudoname:artist.username});
-                    });
-                    if (req.user)
-                        connection.query('SELECT `id` FROM `playlist_likes` WHERE `playlistID` = ? AND `userLogin` = ?',[row.id,req.user.login],(error,likes)=>
-                        {
-                            if (error)
-                            {
-                                response.status(400,error,res);
-                            }
-                            else
-                            {
-                                if (likes.length>0) row.liked = true;
-                                else row.liked = false;
-                                response.status(200,row,res);
-                            }
-                        });
-                    else
-                    {
-                        row.liked = false;
-                        response.status(200,row,res);
-                    }
-                }
-            })
-        }
+                    if (error) return response.status(400,error,res);
+                    row.liked = (likes.length>0)
+                    return response.status(200,row,res);
+                });
+            else
+            {
+                row.liked = false;
+                return response.status(200,row,res);
+            }
+        })
     })
 }
 
@@ -73,14 +46,8 @@ exports.getSongs = (req,res) =>
 {
     connection.query("SELECT `id`, ROW_NUMBER() OVER(PARTITION BY null ORDER BY `pos` ASC) AS `pos` FROM `view_playlist_songs` WHERE `playlistID` = ? ORDER BY `view_playlist_songs`.`pos`",[req.params.id],(error,rows,fields)=>
     {
-        if (error)
-        {
-            response.status(400,error,res);
-        }
-        else
-        {
-            response.status(200,rows,res);
-        }
+        if (error) return response.status(400,error,res);
+        return response.status(200,rows,res);
     })
 }
 
@@ -88,56 +55,34 @@ exports.getCover = (req,res) =>
 {
     connection.query("SELECT `coversrc` FROM `playlists` WHERE `id` = ?",[req.params.id],(error,rows,fields)=>
     {
-        if (error)
+        if (error) return response.status(400,error,res);
+        if (rows.length<1) return response.status(404,'Playlist not found',res);
+        const row = rows[0];
+        res.sendFile("images/covers/"+row.coversrc,{root: '.'}, function (error)
         {
-            response.status(400,error,res);
-        }
-        else if (rows.length<1)
-        {
-            response.status(404,'Playlist not found',res);
-        }
-        else
-        {
-            const row = rows[0];
-            res.sendFile("images/covers/"+row.coversrc,{root: '.'}, function (error)
-            {
-                if (error)
-                {
-                  response.status(error.status,error,res);
-                }
-            });
-        }
+            if (error) return response.status(error.status,error,res);
+        });
     })
 }
 
 exports.postLike = (req,res) =>
 {
+    if (!req.user?.login) return response.status(401,"no auth",res);
     const sql = 'INSERT INTO `playlist_likes`(`userLogin`,`playlistID`,`time`) VALUES (?,?,?)';
     connection.query(sql,[req.user.login,req.params.id,new Date().toISOString().slice(0, 19).replace('T', ' ')],(error,results)=>
     {
-        if (error)
-        {
-            response.status(400,error,res);
-        }
-        else
-        {
-            response.status(200,'Playlist liked',res);
-        }
+        if (error) return response.status(400,error,res);
+        return response.status(201,'Playlist liked',res);
     })
 }
 
 exports.deleteLike = (req,res) =>
 {
+    if (!req.user?.login) return response.status(401,"no auth",res);
     const sql = 'DELETE FROM `playlist_likes` WHERE`userLogin`=? AND `playlistID`=?';
     connection.query(sql,[req.user.login,req.params.id],(error,results)=>
     {
-        if (error)
-        {
-            response.status(400,error,res);
-        }
-        else
-        {
-            response.status(200,'Playlist disliked',res);
-        }
+        if (error) return response.status(400,error,res);
+        return response.status(201,'Playlist disliked',res);
     })
 }
