+++ controller/songController.js
@@ -5,14 +5,8 @@ exports.getAll = (req,res) =>
 {
     connection.query("SELECT `id`,ROW_NUMBER() OVER(PARTITION BY null ORDER BY `songs`.`created_at` DESC) AS `pos` FROM `songs`",(error,rows,fields)=>
     {
-        if (error)
-        {
-            response.status(400,error,res);
-        }
-        else
-        {
-            response.status(200,rows,res);
-        }
+        if (error) return response.status(400,error,res);
+        return response.status(200,rows,res);
     })
 }
 
@@ -20,52 +14,30 @@ exports.getByID = (req,res) =>
 {
     connection.query("SELECT * FROM `view_song` WHERE `id` = ?",[req.params.id],(error,rows,fields)=>
     {
-        if (error)
+        if (error) return response.status(400,error,res);
+        if (rows.length<1) return response.status(404,'Song not found',res);
+        const row = rows[0];
+        row.artists = [];
+        connection.query('SELECT `login`,`username`,`pseudoname` FROM `view_song_artists` WHERE `songID` = ? ORDER BY `view_song_artists`.`artistSongPosition`',[row.id],(error,artists)=>
         {
-            response.status(400,error,res);
-        }
-        else if (rows.length<1)
-        {
-            response.status(404,'Song not found',res);
-        }
-        else
-        {
-            const row = rows[0];
-            row.artists = [];
-            connection.query('SELECT `login`,`username`,`pseudoname` FROM `view_song_artists` WHERE `songID` = ? ORDER BY `view_song_artists`.`artistSongPosition`',[row.id],(error,artists)=>
+            if (error) return response.status(400,error,res);
+            artists.forEach(artist=>
             {
-                if (error)
+                row.artists.push({login:artist.login,name:artist.pseudoname || artist.username || artist.login});
+            });
+            if (req.user)
+                connection.query('SELECT `id` FROM `song_likes` WHERE `songID` = ? AND `userLogin` = ?',[row.id,req.user.login],(error,likes)=>
                 {
-                    response.status(400,error,res);
-                }
-                else
-                {
-                    artists.forEach(artist=>
-                    {
-                        row.artists.push({login:artist.login,name:artist.pseudoname || artist.username || artist.login});
-                    });
-                    if (req.user)
-                        connection.query('SELECT `id` FROM `song_likes` WHERE `songID` = ? AND `userLogin` = ?',[row.id,req.user.login],(error,likes)=>
-                        {
-                            if (error)
-                            {
-                                response.status(400,error,res);
-                            }
-                            else
-                            {
-                                if (likes.length>0) row.liked = true;
-                                else row.liked = false;
-                                response.status(200,row,res);
-                            }
-                        });
-                    else
-                    {
-                        row.liked = false;
-                        response.status(200,row,res);
-                    }
-                }
-            })
-        }
+                    if (error) return response.status(400,error,res);
+                    row.liked = (likes.length>0);
+                    return response.status(200,row,res);
+                });
+            else
+            {
+                row.liked = false;
+                return response.status(200,row,res);
+            }
+        })
     })
 }
 
@@ -73,22 +45,13 @@ exports.getAudio = (req,res) =>
 {
     connection.query("SELECT `audiosrc` FROM `songs` WHERE `id` = ?",[req.params.id],(error,rows,fields)=>
     {
-        if (error)
+        if (error) return response.status(400,error,res);
+        if (rows.length<1) return response.status(404,'Song not found',res);
+        const row = rows[0];
+        res.sendFile("audio/"+row.audiosrc,{root: '.'}, (error)=>
         {
-            response.status(400,error,res);
-        }
-        else if (rows.length<1)
-        {
-            response.status(404,'Song not found',res);
-        }
-        else
-        {
-            const row = rows[0];
-            res.sendFile("audio/"+row.audiosrc,{root: '.'}, (error)=>
-            {
-                if (error) response.status(error.status,error,res);
-            });
-        }
+            if (error) return response.status(error.status,error,res);
+        });
     })
 }
 
@@ -96,25 +59,13 @@ exports.getCover = (req,res) =>
 {
     connection.query("SELECT `coversrc` FROM `songs` WHERE `id` = ?",[req.params.id],(error,rows,fields)=>
     {
-        if (error)
+        if (error) return response.status(400,error,res);
+        if (rows.length<1) return response.status(404,'Song not found',res);
+        const row = rows[0];
+        res.sendFile("images/covers/"+row.coversrc,{root: '.'}, function (error)
         {
-            response.status(400,error,res);
-        }
-        else if (rows.length<1)
-        {
-            response.status(404,'Song not found',res);
-        }
-        else
-        {
-            const row = rows[0];
-            res.sendFile("images/covers/"+row.coversrc,{root: '.'}, function (error)
-            {
-                if (error)
-                {
-                  response.status(error.status,error,res);
-                }
-            });
-        }
+            if (error) return response.status(error.status,error,res);
+        });
     })
 }
 
@@ -122,14 +73,8 @@ exports.getLikes = (req,res) =>
 {
     connection.query("SELECT `userLogin` as `login` FROM `song_likes` WHERE `songID` = ? ORDER BY `time` DESC",[req.params.id],(error,rows,fields)=>
     {
-        if (error)
-        {
-            response.status(400,error,res);
-        }
-        else
-        {
-            response.status(200,rows,res);
-        }
+        if (error) return response.status(400,error,res);
+        return response.status(200,rows,res);
     })
 }
 
@@ -137,45 +82,29 @@ exports.getPlaylists = (req,res) =>
 {
     connection.query("SELECT DISTINCT `playlistID` as `id` FROM `playlist_songs` WHERE `songID` = ?",[req.params.id],(error,rows,fields)=>
     {
-        if (error)
-        {
-            response.status(400,error,res);
-        }
-        else
-        {
-            response.status(200,rows,res);
-        }
+        if (error) return response.status(400,error,res);
+        return response.status(200,rows,res);
     })
 }
 
 exports.postLike = (req,res) =>
 {
+    if (!req.user?.login) return response.status(401,"no auth",res);
     const sql = 'INSERT INTO `song_likes`(`userLogin`,`songID`,`time`) VALUES (?,?,?)';
     connection.query(sql,[req.user.login,req.params.id,new Date().toISOString().slice(0, 19).replace('T', ' ')],(error,results)=>
     {
-        if (error)
-        {
-            response.status(400,error,res);
-        }
-        else
-        {
-            response.status(200,'Song liked',res);
-        }
+        if (error) return response.status(400,error,res);
+        return response.status(201,'Song liked',res);
     })
 }
 
 exports.deleteLike = (req,res) =>
 {
+    if (!req.user?.login) return response.status(401,"no auth",res);
     const sql = 'DELETE FROM `song_likes` WHERE`userLogin`=? AND `songID`=?';
     connection.query(sql,[req.user.login,req.params.id],(error,results)=>
     {
-        if (error)
-        {
-            response.status(400,error,res);
-        }
-        else
-        {
-            response.status(200,'Song disliked',res);
-        }
+        if (error) return response.status(400,error,res);
+        return response.status(201,'Song disliked',res);
     })
 }
